@@ -1,4 +1,2 @@
-// this require is used in dev only anyways
-require('time-require')
-
-require('./bin')
+import 'time-require'
+import './bin.js'
