@@ -1,7 +1,7 @@
 import createDebug from 'debug'
 import ora from 'ora'
-import type { ActionResult } from '../types'
-import createResult from './result'
+import type { ActionResult } from '../types.js'
+import createResult from './result.js'
 const debug = createDebug('hygen:ops:shell')
 
 const notEmpty = (x) => x && x.length > 0
